backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
qdrant-client
langchain
langchain-community
pymupdf
requests
fastembed
spacy
redis
pyyaml
